requirements.txt
deta==1.1.0
geopy==2.3.0
langchain==0.0.205
loguru==0.7.0
meteostat==1.6.5
numpy==1.25.0
openai==0.27.8
openpyxl==3.1.2
pandas==2.0.2
pandas-stubs
plotly==5.15.0
polars==0.18.3
PyGithub==1.58.2
pytest==7.3.2
python-dotenv==1.0.0
scipy==1.10.1
sentry-sdk==1.25.1
streamlit-authenticator==0.2.2
streamlit-lottie==0.0.5
streamlit==1.23.1
xlsx2csv==0.8.1
XlsxWriter==3.1.2
